hw/block/nvme: refactor nvme_process_sq() specifically
This patch added nothing but a nvme_process_sqe() which will handle
fetched one(sqe) from the sq.  During that time, Cleaned up the code of
retrieving the NvmeRequest instance for the command fetched.

These two functions are going to be taken for the different use-cases
such as FUSED operations from normal use-cases.

This patch has no functional changes.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>