hw/block/nvme: support for FUSED operation (compare & write)
This patch introduces support for FUSED operation for the pair of
	COMPARE and WRITE

It has added FUSED check code to nvme_process_sq() which is one of the
main job for the device, but with the unlikely() to make it harmless for
the performance.

This patch has stopped the sq timers to make sure the device handles two
commands(FUSE) in an atomic context.  After that, it will unquiesce the
timers to run again.

Limitation:
  We don't actually detect the submission queue tail doorbell update for
more than 1 entry for FUSED operation in the stage of the doorbell
udpate (mmio) because this actually means that we need some burst
arbitration feature for sq entries pushed.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>